Renamed folder to 'testing'
